// src/services/types/platform/screen_annotation.ts
export type ScreenAnnotationApiResponse = {
  _id: string
  __v: number
  screen_annotation: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
    }
    section2_tab1: {
      image1: string
      title1_en: string
      title1_fa: string
      video1: string
    }
    section2_tab2: {
      image1: string
      title1_en: string
      title1_fa: string
      video1: string
    }
    section3: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      image1: string
    }
    section4: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
      image1: string
    }
    section5: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
}